Move all interfaces to Api folder
