+++ Api/ConfiguratorAdapterInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CtiDigital\Configurator\Model;
+namespace CtiDigital\Configurator\Api;
 
 interface ConfiguratorAdapterInterface
 {
