+++ Component/ApiIntegrations.php
@@ -4,7 +4,7 @@ namespace CtiDigital\Configurator\Component;
 use Magento\Framework\ObjectManagerInterface;
 use Magento\Integration\Model\IntegrationFactory;
 use Magento\Integration\Model\Oauth\TokenFactory;
-use CtiDigital\Configurator\Model\LoggerInterface;
+use CtiDigital\Configurator\Api\LoggerInterface;
 use Magento\Integration\Model\AuthorizationService;
 use Magento\Integration\Api\IntegrationServiceInterface;
 use CtiDigital\Configurator\Model\Exception\ComponentException;
