+++ Component/Blocks.php
@@ -3,7 +3,7 @@
 namespace CtiDigital\Configurator\Component;
 
 use CtiDigital\Configurator\Model\Exception\ComponentException;
-use CtiDigital\Configurator\Model\LoggerInterface;
+use CtiDigital\Configurator\Api\LoggerInterface;
 use Magento\Cms\Api\Data\BlockInterfaceFactory;
 use Magento\Framework\ObjectManagerInterface;
