+++ Component/CustomerGroups.php
@@ -5,7 +5,7 @@ use Symfony\Component\Yaml\Yaml;
 use Magento\Customer\Model\GroupFactory;
 use Magento\Tax\Model\ClassModelFactory;
 use Magento\Framework\ObjectManagerInterface;
-use CtiDigital\Configurator\Model\LoggerInterface;
+use CtiDigital\Configurator\Api\LoggerInterface;
 use CtiDigital\Configurator\Model\Exception\ComponentException;
 
 class CustomerGroups extends YamlComponentAbstract
