+++ Component/Customers.php
@@ -1,7 +1,7 @@
 <?php
 namespace CtiDigital\Configurator\Component;
 
-use CtiDigital\Configurator\Model\LoggerInterface;
+use CtiDigital\Configurator\Api\LoggerInterface;
 use CtiDigital\Configurator\Model\Exception\ComponentException;
 use Magento\Framework\ObjectManagerInterface;
 use FireGento\FastSimpleImport\Model\ImporterFactory;
