+++ Component/Rewrites.php
@@ -5,7 +5,7 @@ namespace CtiDigital\Configurator\Component;
 use Magento\Framework\ObjectManagerInterface;
 use Magento\UrlRewrite\Model\UrlRewriteFactory;
 use Magento\UrlRewrite\Model\UrlPersistInterface;
-use CtiDigital\Configurator\Model\LoggerInterface;
+use CtiDigital\Configurator\Api\LoggerInterface;
 use CtiDigital\Configurator\Model\Exception\ComponentException;
 
 class Rewrites extends CsvComponentAbstract
