+++ Model/Processor.php
@@ -2,8 +2,9 @@
 
 namespace CtiDigital\Configurator\Model;
 
+use CtiDigital\Configurator\Api\LoggerInterface;
 use CtiDigital\Configurator\Component\ComponentAbstract;
-use CtiDigital\Configurator\Model\Configurator\ConfigInterface;
+use CtiDigital\Configurator\Api\ConfigInterface;
 use CtiDigital\Configurator\Model\Exception\ComponentException;
 use Magento\Framework\ObjectManagerInterface;
 use Symfony\Component\Yaml\Parser;
@@ -12,7 +13,6 @@ use Magento\Framework\App\Area;
 
 class Processor
 {
-
     /**
      * @var string
      */
