+++ etc/di.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0"?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
 
-    <preference for="CtiDigital\Configurator\Model\ConfiguratorAdapterInterface"
-                type="CtiDigital\Configurator\Model\ConfiguratorAdapterInterface\Proxy"/>
-    <preference for="CtiDigital\Configurator\Model\Configurator\ConfigInterface"
+    <preference for="CtiDigital\Configurator\Api\ConfiguratorAdapterInterface"
+                type="CtiDigital\Configurator\Api\ConfiguratorAdapterInterface\Proxy"/>
+    <preference for="CtiDigital\Configurator\Api\ConfigInterface"
                 type="CtiDigital\Configurator\Model\Configurator\Config" />
-    <preference for="CtiDigital\Configurator\Model\LoggerInterface"
+    <preference for="CtiDigital\Configurator\Api\LoggerInterface"
                 type="CtiDigital\Configurator\Model\Logging" />
 
     <type name="Magento\Framework\Console\CommandList">
